refactor(wdotool): extract dispatch into a library with injected I/O
Splits the binary into a wdotool lib and a thin main. dispatch() now
takes a DispatchCtx with stdout/stderr writers and returns a structured
ExitCode instead of calling process::exit. The binary calls
process::exit on non-zero; integration tests in tests/ can drive the
same dispatch path against any Backend impl and assert on captured
output without spawning a subprocess.

No behavior change for the CLI. All 36 existing unit tests still pass.
This sets up a mock-backend integration test suite (next commit) and
the headless wlroots harness later.
